+++ rust/ast_utils/src/lib.rs
@@ -1,5 +1,8 @@
 pub mod converters;
 
+// simple re-usable helpers for unit tests
+mod unit_test_helpers;
+
 #[cfg(test)]
 mod tests {
     #[test]
